+++ src/models/ad-config.ts
@@ -1,5 +1,6 @@
 import { AdConfigData } from "../types/ad-config";
 import { config } from "../api";
+import { rethrowPlatformError } from "../utils/error-handler";
 
 /** @hidden */
 export default class AdConfig {
@@ -64,12 +65,14 @@ export default class AdConfig {
                 for (let i = 0; i < adUnits.length; i++) {
                     if (adUnits[i].type === "INTERSTITIAL") {
                         this._current.interstitialId = adUnits[i].id;
-                    }
-                    else if (adUnits[i].type === "REWARDED_VIDEO") {
+                    } else if (adUnits[i].type === "REWARDED_VIDEO") {
                         this._current.rewardedId = adUnits[i].id;
                     }
                 }
-            });
+            })
+                .catch((e: any) => {
+                    rethrowPlatformError(e, "setLinkViberAdUnitIds()");
+                });
         }
     }
 
@@ -89,26 +92,22 @@ export default class AdConfig {
     private setFacebookAdUnitIds(): void {
         if ((window as any).wortalGame) {
             (window as any).wortalGame.getAdUnitIDAsync().then((adUnits: any) => {
-                if (adUnits == null || undefined) {
+                if ((adUnits == null || undefined) || (adUnits.ads == null || undefined)) {
                     console.error("[Wortal] Failed to retrieve ad units.");
                     return;
                 }
-                if (adUnits.ads == null || undefined) {
-                    console.error("[Wortal] Failed to retrieve ad units.");
-                    return;
-                }
-                console.log("[Wortal] AdUnit IDs returned: \n" + adUnits.ads);
+                console.log("[Wortal] AdUnit IDs returned: \n" + adUnits);
                 for (let i = 0; i < adUnits.ads.length; i++) {
                     if (adUnits.ads[i].display_format === "interstitial") {
                         this._current.interstitialId = adUnits.ads[i].placement_id;
-                    } else if (adUnits.ads[i].display_format === "rewarded") {
+                    } else if (adUnits.ads[i].display_format === "rewarded_video") {
                         this._current.rewardedId = adUnits.ads[i].placement_id;
                     }
                 }
             })
-            .catch((e: any) => {
-                throw Error(e);
-            });
+                .catch((e: any) => {
+                    rethrowPlatformError(e, "setFacebookAdUnitIds()");
+                });
         }
     }
 }